zingen::asm

Struct Assembler

pub struct Assembler {
    buffer: Buffer,
    gas: u128,
    pub mp: usize,
    pub sp: u16,
}

Low level assembler implementation for EVM.

Fields
buffer: Buffer

Buffer of the assembler.

gas: u128

Gas counter.

This is used to calculate the gas cost of the generated code.

TODO: use a more precise type, eq u256. (issue-20)

mp: usize

Memory pointer for byte offset.

sp: u16

Stack pointer, maximum MAX_STACK_SIZE items.

Implementations

impl Assembler

pub fn buffer(&self) -> &[u8]

Buffer of the assembler.

pub fn buffer_mut(&mut self) -> &mut Buffer

Mutable buffer of the assembler.

pub fn increment_gas(&mut self, gas: u128)

Increment the gas counter.

TODO: use number bigger than u256 for throwing proper errors. (#21)

pub fn increment_sp(&mut self, items: u16) -> Result<()>

Increment stack pointer

pub fn decrement_sp(&mut self, items: u16) -> Result<()>

Decrement stack pointer

pub fn increment_mp(&mut self, offset: usize) -> Result<()>

Increment memory pointer

pub fn decrement_mp(&mut self, offset: usize) -> Result<()>

Decrement memory pointer

pub fn emit(&mut self, byte: u8)

Emit a byte.

pub fn emitn(&mut self, bytes: &[u8])

Emit n bytes.

pub fn emit_op(&mut self, opcode: OpCode) -> Result<()>

Emit a single opcode.

Mock the stack input and output for checking the stack usages.

impl Assembler

Basic instruction implementations

pub fn _stop(&mut self) -> Result<()>

Emit STOP

pub fn _add(&mut self) -> Result<()>

Emit ADD

pub fn _mul(&mut self) -> Result<()>

Emit MUL

pub fn _sub(&mut self) -> Result<()>

Emit SUB

pub fn _div(&mut self) -> Result<()>

Emit DIV

pub fn _sdiv(&mut self) -> Result<()>

Emit SDIV

pub fn _mod(&mut self) -> Result<()>

Emit MOD

pub fn _smod(&mut self) -> Result<()>

Emit SMOD

pub fn _addmod(&mut self) -> Result<()>

Emit ADDMOD

pub fn _mulmod(&mut self) -> Result<()>

Emit MULMOD

pub fn _exp(&mut self) -> Result<()>

Emit EXP

pub fn _signextend(&mut self) -> Result<()>

Emit SIGNEXTEND

pub fn _lt(&mut self) -> Result<()>

Emit LT

pub fn _gt(&mut self) -> Result<()>

Emit GT

pub fn _slt(&mut self) -> Result<()>

Emit SLT

pub fn _sgt(&mut self) -> Result<()>

Emit SGT

pub fn _eq(&mut self) -> Result<()>

Emit EQ

pub fn _iszero(&mut self) -> Result<()>

Emit ISZERO

pub fn _and(&mut self) -> Result<()>

Emit AND

pub fn _or(&mut self) -> Result<()>

Emit OR

pub fn _xor(&mut self) -> Result<()>

Emit XOR

pub fn _not(&mut self) -> Result<()>

Emit NOT

pub fn _byte(&mut self) -> Result<()>

Emit BYTE

pub fn _shl(&mut self) -> Result<()>

Emit SHL

pub fn _shr(&mut self) -> Result<()>

Emit SHR

pub fn _sar(&mut self) -> Result<()>

Emit SAR

pub fn _keccak256(&mut self) -> Result<()>

Emit KECCAK256

pub fn _address(&mut self) -> Result<()>

Emit ADDRESS

pub fn _balance(&mut self) -> Result<()>

Emit BALANCE

pub fn _origin(&mut self) -> Result<()>

Emit ORIGIN

pub fn _caller(&mut self) -> Result<()>

Emit CALLER

pub fn _callvalue(&mut self) -> Result<()>

Emit CALLVALUE

pub fn _calldataload(&mut self) -> Result<()>

Emit CALLDATALOAD

pub fn _calldatasize(&mut self) -> Result<()>

Emit CALLDATASIZE

pub fn _calldatacopy(&mut self) -> Result<()>

Emit CALLDATACOPY

pub fn _codesize(&mut self) -> Result<()>

Emit CODESIZE

pub fn _codecopy(&mut self) -> Result<()>

Emit CODECOPY

pub fn _gasprice(&mut self) -> Result<()>

Emit GASPRICE

pub fn _extcodesize(&mut self) -> Result<()>

Emit EXTCODESIZE

pub fn _extcodecopy(&mut self) -> Result<()>

Emit EXTCODECOPY

pub fn _returndatasize(&mut self) -> Result<()>

Emit RETURNDATASIZE

pub fn _returndatacopy(&mut self) -> Result<()>

Emit RETURNDATACOPY

pub fn _extcodehash(&mut self) -> Result<()>

Emit EXTCODEHASH

pub fn _blockhash(&mut self) -> Result<()>

Emit BLOCKHASH

pub fn _coinbase(&mut self) -> Result<()>

Emit COINBASE

pub fn _timestamp(&mut self) -> Result<()>

Emit TIMESTAMP

pub fn _number(&mut self) -> Result<()>

Emit NUMBER

pub fn _difficulty(&mut self) -> Result<()>

Emit DIFFICULTY

pub fn _gaslimit(&mut self) -> Result<()>

Emit GASLIMIT

pub fn _chainid(&mut self) -> Result<()>

Emit CHAINID

pub fn _selfbalance(&mut self) -> Result<()>

Emit SELFBALANCE

pub fn _basefee(&mut self) -> Result<()>

Emit BASEFEE

pub fn _blobhash(&mut self) -> Result<()>

Emit BLOBHASH

pub fn _blobbasefee(&mut self) -> Result<()>

Emit BLOBBASEFEE

pub fn _pop(&mut self) -> Result<()>

Emit POP

pub fn _mload(&mut self) -> Result<()>

Emit MLOAD

pub fn _mstore(&mut self) -> Result<()>

Emit MSTORE

pub fn _mstore8(&mut self) -> Result<()>

Emit MSTORE8

pub fn _sload(&mut self) -> Result<()>

Emit SLOAD

pub fn _sstore(&mut self) -> Result<()>

Emit SSTORE

pub fn _jump(&mut self) -> Result<()>

Emit JUMP

pub fn _jumpi(&mut self) -> Result<()>

Emit JUMPI

pub fn _pc(&mut self) -> Result<()>

Emit PC

pub fn _msize(&mut self) -> Result<()>

Emit MSIZE

pub fn _gas(&mut self) -> Result<()>

Emit GAS

pub fn _jumpdest(&mut self) -> Result<()>

Emit JUMPDEST

pub fn _tload(&mut self) -> Result<()>

Emit TLOAD

pub fn _tstore(&mut self) -> Result<()>

Emit TSTORE

pub fn _mcopy(&mut self) -> Result<()>

Emit MCOPY

pub fn _push0(&mut self) -> Result<()>

Emit PUSH0

pub fn _push1(&mut self) -> Result<()>

Emit PUSH1

pub fn _push2(&mut self) -> Result<()>

Emit PUSH2

pub fn _push3(&mut self) -> Result<()>

Emit PUSH3

pub fn _push4(&mut self) -> Result<()>

Emit PUSH4

pub fn _push5(&mut self) -> Result<()>

Emit PUSH5

pub fn _push6(&mut self) -> Result<()>

Emit PUSH6

pub fn _push7(&mut self) -> Result<()>

Emit PUSH7

pub fn _push8(&mut self) -> Result<()>

Emit PUSH8

pub fn _push9(&mut self) -> Result<()>

Emit PUSH9

pub fn _push10(&mut self) -> Result<()>

Emit PUSH10

pub fn _push11(&mut self) -> Result<()>

Emit PUSH11

pub fn _push12(&mut self) -> Result<()>

Emit PUSH12

pub fn _push13(&mut self) -> Result<()>

Emit PUSH13

pub fn _push14(&mut self) -> Result<()>

Emit PUSH14

pub fn _push15(&mut self) -> Result<()>

Emit PUSH15

pub fn _push16(&mut self) -> Result<()>

Emit PUSH16

pub fn _push17(&mut self) -> Result<()>

Emit PUSH17

pub fn _push18(&mut self) -> Result<()>

Emit PUSH18

pub fn _push19(&mut self) -> Result<()>

Emit PUSH19

pub fn _push20(&mut self) -> Result<()>

Emit PUSH20

pub fn _push21(&mut self) -> Result<()>

Emit PUSH21

pub fn _push22(&mut self) -> Result<()>

Emit PUSH22

pub fn _push23(&mut self) -> Result<()>

Emit PUSH23

pub fn _push24(&mut self) -> Result<()>

Emit PUSH24

pub fn _push25(&mut self) -> Result<()>

Emit PUSH25

pub fn _push26(&mut self) -> Result<()>

Emit PUSH26

pub fn _push27(&mut self) -> Result<()>

Emit PUSH27

pub fn _push28(&mut self) -> Result<()>

Emit PUSH28

pub fn _push29(&mut self) -> Result<()>

Emit PUSH29

pub fn _push30(&mut self) -> Result<()>

Emit PUSH30

pub fn _push31(&mut self) -> Result<()>

Emit PUSH31

pub fn _push32(&mut self) -> Result<()>

Emit PUSH32

pub fn _dup1(&mut self) -> Result<()>

Emit DUP1

pub fn _dup2(&mut self) -> Result<()>

Emit DUP2

pub fn _dup3(&mut self) -> Result<()>

Emit DUP3

pub fn _dup4(&mut self) -> Result<()>

Emit DUP4

pub fn _dup5(&mut self) -> Result<()>

Emit DUP5

pub fn _dup6(&mut self) -> Result<()>

Emit DUP6

pub fn _dup7(&mut self) -> Result<()>

Emit DUP7

pub fn _dup8(&mut self) -> Result<()>

Emit DUP8

pub fn _dup9(&mut self) -> Result<()>

Emit DUP9

pub fn _dup10(&mut self) -> Result<()>

Emit DUP10

pub fn _dup11(&mut self) -> Result<()>

Emit DUP11

pub fn _dup12(&mut self) -> Result<()>

Emit DUP12

pub fn _dup13(&mut self) -> Result<()>

Emit DUP13

pub fn _dup14(&mut self) -> Result<()>

Emit DUP14

pub fn _dup15(&mut self) -> Result<()>

Emit DUP15

pub fn _dup16(&mut self) -> Result<()>

Emit DUP16

pub fn _swap1(&mut self) -> Result<()>

Emit SWAP1

pub fn _swap2(&mut self) -> Result<()>

Emit SWAP2

pub fn _swap3(&mut self) -> Result<()>

Emit SWAP3

pub fn _swap4(&mut self) -> Result<()>

Emit SWAP4

pub fn _swap5(&mut self) -> Result<()>

Emit SWAP5

pub fn _swap6(&mut self) -> Result<()>

Emit SWAP6

pub fn _swap7(&mut self) -> Result<()>

Emit SWAP7

pub fn _swap8(&mut self) -> Result<()>

Emit SWAP8

pub fn _swap9(&mut self) -> Result<()>

Emit SWAP9

pub fn _swap10(&mut self) -> Result<()>

Emit SWAP10

pub fn _swap11(&mut self) -> Result<()>

Emit SWAP11

pub fn _swap12(&mut self) -> Result<()>

Emit SWAP12

pub fn _swap13(&mut self) -> Result<()>

Emit SWAP13

pub fn _swap14(&mut self) -> Result<()>

Emit SWAP14

pub fn _swap15(&mut self) -> Result<()>

Emit SWAP15

pub fn _swap16(&mut self) -> Result<()>

Emit SWAP16

pub fn _log0(&mut self) -> Result<()>

Emit LOG0

pub fn _log1(&mut self) -> Result<()>

Emit LOG1

pub fn _log2(&mut self) -> Result<()>

Emit LOG2

pub fn _log3(&mut self) -> Result<()>

Emit LOG3

pub fn _log4(&mut self) -> Result<()>

Emit LOG4

pub fn _create(&mut self) -> Result<()>

Emit CREATE

pub fn _call(&mut self) -> Result<()>

Emit CALL

pub fn _callcode(&mut self) -> Result<()>

Emit CALLCODE

pub fn _return(&mut self) -> Result<()>

Emit RETURN

pub fn _delegatecall(&mut self) -> Result<()>

Emit DELEGATECALL

pub fn _create2(&mut self) -> Result<()>

Emit CREATE2

pub fn _staticcall(&mut self) -> Result<()>

Emit STATICCALL

pub fn _revert(&mut self) -> Result<()>

Emit REVERT

pub fn _invalid(&mut self) -> Result<()>

Emit INVALID

pub fn _selfdestruct(&mut self) -> Result<()>

Emit SELFDESTRUCT