zingen::result

Enum Error

pub enum Error {

    Anyhow(Error),
    Abi(Error),
    BinaryReader(BinaryReaderError),
    BufferOverflow(usize),
    ControlStackUnderflow,
    DataNotFound(i32, usize),
    DuplicateFunc(u32),
    DuplicateJump(u16),
    ExtFuncNotFound,
    FuncNotFound(u32),
    FuncNotImported(String),
    HostFuncNotFound(String, String),
    ImportedFuncNotFound(u32),
    InvalidElseBlock(u16),
    InvalidFunctionSignature,
    InvalidLocalIndex(usize),
    InvalidMP(u8),
    InvalidPC(usize),
    InvalidDataOffset(i32),
    InvalidDataSize(usize),
    InvalidDepth(usize),
    InvalidSelector,
    LabelMismatch,
    LocalIndexOutOfRange,
    LocalNotOnStack(usize),
    MemoryOutOfBounds,
    SelectorNotFound,
    StackIndexOutOfRange(u16),
    StackOverflow(u16, u16),
    StackUnderflow(u16, u16),
    StackNotBalanced(u16),
    UnsupportedHostFunc(HostFunc),

}

Codegen error

Variants

Anyhow(Error)

Any error

Abi(Error)

Failed to parse function ABI.

BinaryReader(BinaryReaderError)

Failed to parse WASM with binary reader.

BufferOverflow(usize)

Failed to push more data to the buffer.

ControlStackUnderflow

Failed to pop control stack frame.

DataNotFound(i32, usize)

Data not found in data section.

DuplicateFunc(u32)

Failed to register program counter to function index.

DuplicateJump(u16)

Failed to merge jump table.

ExtFuncNotFound

Failed to find ext function index in jump table.

FuncNotFound(u32)

Failed to find function index in jump table.

FuncNotImported(String)

Failed to find function index in jump table.

HostFuncNotFound(String, String)

Failed to find host function in compiler.

ImportedFuncNotFound(u32)

Failed to find imported function by index in jump table.

InvalidElseBlock(u16)

Failed to mark else block for if block.

InvalidFunctionSignature

Failed parse function signature.

InvalidLocalIndex(usize)

Failed to get local with given index.

InvalidMP(u8)

Failed to get the offset of the given memory pointer.

InvalidPC(usize)

Failed to construct program counter for jump.

InvalidDataOffset(i32)

Failed to get data from the provided offset.

InvalidDataSize(usize)

Failed to get data from the provided offset.

InvalidDepth(usize)

Failed to get frame info of the given depth.

InvalidSelector

Failed to parse function selector.

LabelMismatch

Failed to patch jump destination.

LocalIndexOutOfRange

Failed to define local variable since the index is out of range.

LocalNotOnStack(usize)

Failed to get local variables.

MemoryOutOfBounds

Failed to index data on memory.

SelectorNotFound

Failed to find function selectors.0

StackIndexOutOfRange(u16)

Failed to index data on stack.

StackOverflow(u16, u16)

Failed to increment stack pointer.

StackUnderflow(u16, u16)

Failed to decrement stack pointer.

StackNotBalanced(u16)

Failed to pop stack.

UnsupportedHostFunc(HostFunc)

Failed to queue host functions.