Struct zingen::Assembler

pub struct Assembler {
    buffer: Buffer,
    gas: u128,
    pub mp: usize,
    pub sp: u8,
}

Low level assembler implementation for EVM.

Fields
buffer: Buffer

Buffer of the assembler.

gas: u128

Gas counter.

This is used to calculate the gas cost of the generated code.

TODO: use a more precise type, eq u256. (issue-20)

mp: usize

Memory pointer for byte offset.

sp: u8

Stack pointer, maximum 12 items.

Implementations

impl Assembler

pub fn buffer(&self) -> &[u8]

Buffer of the assembler.

pub fn buffer_mut(&mut self) -> &mut Buffer

Mutable buffer of the assembler.

pub fn increment_gas(&mut self, gas: u128)

Increment the gas counter.

TODO: use number bigger than u256 for throwing proper errors. (#21)

pub fn increment_sp(&mut self, items: u8) -> Result<()>

Increment stack pointer

pub fn decrement_sp(&mut self, items: u8) -> Result<()>

Decrement stack pointer

pub fn increment_mp(&mut self, offset: usize) -> Result<()>

Increment memory pointer

pub fn decrement_mp(&mut self, offset: usize) -> Result<()>

Decrement memory pointer

pub fn emit(&mut self, byte: u8)

Emit a byte.

pub fn emitn(&mut self, bytes: &[u8])

Emit n bytes.

pub fn emit_op(&mut self, opcode: OpCode) -> Result<()>

Emit a single opcode.

Mock the stack input and ouput for checking the stack usages.

impl Assembler

Basic instruction implementations

pub fn _stop(&mut self) -> Result<()>

EmitSTOP

pub fn _add(&mut self) -> Result<()>

EmitADD

pub fn _mul(&mut self) -> Result<()>

EmitMUL

pub fn _sub(&mut self) -> Result<()>

EmitSUB

pub fn _div(&mut self) -> Result<()>

EmitDIV

pub fn _sdiv(&mut self) -> Result<()>

EmitSDIV

pub fn _mod(&mut self) -> Result<()>

EmitMOD

pub fn _smod(&mut self) -> Result<()>

EmitSMOD

pub fn _addmod(&mut self) -> Result<()>

EmitADDMOD

pub fn _mulmod(&mut self) -> Result<()>

EmitMULMOD

pub fn _exp(&mut self) -> Result<()>

EmitEXP

pub fn _signextend(&mut self) -> Result<()>

EmitSIGNEXTEND

pub fn _lt(&mut self) -> Result<()>

EmitLT

pub fn _gt(&mut self) -> Result<()>

EmitGT

pub fn _slt(&mut self) -> Result<()>

EmitSLT

pub fn _sgt(&mut self) -> Result<()>

EmitSGT

pub fn _eq(&mut self) -> Result<()>

EmitEQ

pub fn _iszero(&mut self) -> Result<()>

EmitISZERO

pub fn _and(&mut self) -> Result<()>

EmitAND

pub fn _or(&mut self) -> Result<()>

EmitOR

pub fn _xor(&mut self) -> Result<()>

EmitXOR

pub fn _not(&mut self) -> Result<()>

EmitNOT

pub fn _byte(&mut self) -> Result<()>

EmitBYTE

pub fn _shl(&mut self) -> Result<()>

EmitSHL

pub fn _shr(&mut self) -> Result<()>

EmitSHR

pub fn _sar(&mut self) -> Result<()>

EmitSAR

pub fn _sha3(&mut self) -> Result<()>

EmitSHA3

pub fn _address(&mut self) -> Result<()>

EmitADDRESS

pub fn _balance(&mut self) -> Result<()>

EmitBALANCE

pub fn _origin(&mut self) -> Result<()>

EmitORIGIN

pub fn _caller(&mut self) -> Result<()>

EmitCALLER

pub fn _callvalue(&mut self) -> Result<()>

EmitCALLVALUE

pub fn _calldataload(&mut self) -> Result<()>

EmitCALLDATALOAD

pub fn _calldatasize(&mut self) -> Result<()>

EmitCALLDATASIZE

pub fn _calldatacopy(&mut self) -> Result<()>

EmitCALLDATACOPY

pub fn _codesize(&mut self) -> Result<()>

EmitCODESIZE

pub fn _codecopy(&mut self) -> Result<()>

EmitCODECOPY

pub fn _gasprice(&mut self) -> Result<()>

EmitGASPRICE

pub fn _extcodesize(&mut self) -> Result<()>

EmitEXTCODESIZE

pub fn _extcodecopy(&mut self) -> Result<()>

EmitEXTCODECOPY

pub fn _returndatasize(&mut self) -> Result<()>

EmitRETURNDATASIZE

pub fn _returndatacopy(&mut self) -> Result<()>

EmitRETURNDATACOPY

pub fn _blockhash(&mut self) -> Result<()>

EmitBLOCKHASH

pub fn _coinbase(&mut self) -> Result<()>

EmitCOINBASE

pub fn _timestamp(&mut self) -> Result<()>

EmitTIMESTAMP

pub fn _number(&mut self) -> Result<()>

EmitNUMBER

pub fn _difficulty(&mut self) -> Result<()>

EmitDIFFICULTY

pub fn _gaslimit(&mut self) -> Result<()>

EmitGASLIMIT

pub fn _pop(&mut self) -> Result<()>

EmitPOP

pub fn _mload(&mut self) -> Result<()>

EmitMLOAD

pub fn _mstore(&mut self) -> Result<()>

EmitMSTORE

pub fn _mstore8(&mut self) -> Result<()>

EmitMSTORE8

pub fn _sload(&mut self) -> Result<()>

EmitSLOAD

pub fn _sstore(&mut self) -> Result<()>

EmitSSTORE

pub fn _jump(&mut self) -> Result<()>

EmitJUMP

pub fn _jumpi(&mut self) -> Result<()>

EmitJUMPI

pub fn _pc(&mut self) -> Result<()>

EmitPC

pub fn _msize(&mut self) -> Result<()>

EmitMSIZE

pub fn _gas(&mut self) -> Result<()>

EmitGAS

pub fn _jumpdest(&mut self) -> Result<()>

EmitJUMPDEST

pub fn _push0(&mut self) -> Result<()>

EmitPUSH0

pub fn _push1(&mut self) -> Result<()>

EmitPUSH1

pub fn _push2(&mut self) -> Result<()>

EmitPUSH2

pub fn _push3(&mut self) -> Result<()>

EmitPUSH3

pub fn _push4(&mut self) -> Result<()>

EmitPUSH4

pub fn _push5(&mut self) -> Result<()>

EmitPUSH5

pub fn _push6(&mut self) -> Result<()>

EmitPUSH6

pub fn _push7(&mut self) -> Result<()>

EmitPUSH7

pub fn _push8(&mut self) -> Result<()>

EmitPUSH8

pub fn _push9(&mut self) -> Result<()>

EmitPUSH9

pub fn _push10(&mut self) -> Result<()>

EmitPUSH10

pub fn _push11(&mut self) -> Result<()>

EmitPUSH11

pub fn _push12(&mut self) -> Result<()>

EmitPUSH12

pub fn _push13(&mut self) -> Result<()>

EmitPUSH13

pub fn _push14(&mut self) -> Result<()>

EmitPUSH14

pub fn _push15(&mut self) -> Result<()>

EmitPUSH15

pub fn _push16(&mut self) -> Result<()>

EmitPUSH16

pub fn _push17(&mut self) -> Result<()>

EmitPUSH17

pub fn _push18(&mut self) -> Result<()>

EmitPUSH18

pub fn _push19(&mut self) -> Result<()>

EmitPUSH19

pub fn _push20(&mut self) -> Result<()>

EmitPUSH20

pub fn _push21(&mut self) -> Result<()>

EmitPUSH21

pub fn _push22(&mut self) -> Result<()>

EmitPUSH22

pub fn _push23(&mut self) -> Result<()>

EmitPUSH23

pub fn _push24(&mut self) -> Result<()>

EmitPUSH24

pub fn _push25(&mut self) -> Result<()>

EmitPUSH25

pub fn _push26(&mut self) -> Result<()>

EmitPUSH26

pub fn _push27(&mut self) -> Result<()>

EmitPUSH27

pub fn _push28(&mut self) -> Result<()>

EmitPUSH28

pub fn _push29(&mut self) -> Result<()>

EmitPUSH29

pub fn _push30(&mut self) -> Result<()>

EmitPUSH30

pub fn _push31(&mut self) -> Result<()>

EmitPUSH31

pub fn _push32(&mut self) -> Result<()>

EmitPUSH32

pub fn _dup1(&mut self) -> Result<()>

EmitDUP1

pub fn _dup2(&mut self) -> Result<()>

EmitDUP2

pub fn _dup3(&mut self) -> Result<()>

EmitDUP3

pub fn _dup4(&mut self) -> Result<()>

EmitDUP4

pub fn _dup5(&mut self) -> Result<()>

EmitDUP5

pub fn _dup6(&mut self) -> Result<()>

EmitDUP6

pub fn _dup7(&mut self) -> Result<()>

EmitDUP7

pub fn _dup8(&mut self) -> Result<()>

EmitDUP8

pub fn _dup9(&mut self) -> Result<()>

EmitDUP9

pub fn _dup10(&mut self) -> Result<()>

EmitDUP10

pub fn _dup11(&mut self) -> Result<()>

EmitDUP11

pub fn _dup12(&mut self) -> Result<()>

EmitDUP12

pub fn _dup13(&mut self) -> Result<()>

EmitDUP13

pub fn _dup14(&mut self) -> Result<()>

EmitDUP14

pub fn _dup15(&mut self) -> Result<()>

EmitDUP15

pub fn _dup16(&mut self) -> Result<()>

EmitDUP16

pub fn _swap1(&mut self) -> Result<()>

EmitSWAP1

pub fn _swap2(&mut self) -> Result<()>

EmitSWAP2

pub fn _swap3(&mut self) -> Result<()>

EmitSWAP3

pub fn _swap4(&mut self) -> Result<()>

EmitSWAP4

pub fn _swap5(&mut self) -> Result<()>

EmitSWAP5

pub fn _swap6(&mut self) -> Result<()>

EmitSWAP6

pub fn _swap7(&mut self) -> Result<()>

EmitSWAP7

pub fn _swap8(&mut self) -> Result<()>

EmitSWAP8

pub fn _swap9(&mut self) -> Result<()>

EmitSWAP9

pub fn _swap10(&mut self) -> Result<()>

EmitSWAP10

pub fn _swap11(&mut self) -> Result<()>

EmitSWAP11

pub fn _swap12(&mut self) -> Result<()>

EmitSWAP12

pub fn _swap13(&mut self) -> Result<()>

EmitSWAP13

pub fn _swap14(&mut self) -> Result<()>

EmitSWAP14

pub fn _swap15(&mut self) -> Result<()>

EmitSWAP15

pub fn _swap16(&mut self) -> Result<()>

EmitSWAP16

pub fn _log0(&mut self) -> Result<()>

EmitLOG0

pub fn _log1(&mut self) -> Result<()>

EmitLOG1

pub fn _log2(&mut self) -> Result<()>

EmitLOG2

pub fn _log3(&mut self) -> Result<()>

EmitLOG3

pub fn _log4(&mut self) -> Result<()>

EmitLOG4

pub fn _create(&mut self) -> Result<()>

EmitCREATE

pub fn _call(&mut self) -> Result<()>

EmitCALL

pub fn _callcode(&mut self) -> Result<()>

EmitCALLCODE

pub fn _return(&mut self) -> Result<()>

EmitRETURN

pub fn _delegatecall(&mut self) -> Result<()>

EmitDELEGATECALL

pub fn _create2(&mut self) -> Result<()>

EmitCREATE2

pub fn _staticcall(&mut self) -> Result<()>

EmitSTATICCALL

pub fn _revert(&mut self) -> Result<()>

EmitREVERT

pub fn _invalid(&mut self) -> Result<()>

EmitINVALID

pub fn _selfdestruct(&mut self) -> Result<()>

EmitSELFDESTRUCT