Struct zingen::CodeGen

pub struct CodeGen {
    pub(crate) backtrace: Backtrace,
    pub(crate) control: ControlStack,
    pub(crate) dataset: DataSet,
    pub(crate) env: FuncType,
    pub(crate) locals: Locals,
    pub(crate) masm: MacroAssembler,
    pub(crate) table: JumpTable,
    pub(crate) imports: Imports,
    pub(crate) is_main: bool,
}

The code generation abstraction.

Fields
backtrace: Backtrace

The backtrace.

control: ControlStack

Control stack frames.

dataset: DataSet

Control stack frames.

env: FuncType

The function environment.

locals: Locals

The defined locals for a function.

masm: MacroAssembler

The macro assembler.

table: JumpTable

The jump table.

imports: Imports

The imported functions.

is_main: bool

If this function is the main function.

Implementations

impl CodeGen

pub fn new( env: FuncType, dataset: DataSet, imports: Imports, is_main: bool ) -> Result<Self>

Create a new code generator.

pub fn emit_locals( &mut self, locals: &mut LocalsReader<'_>, validator: &mut FuncValidator<ValidatorResources> ) -> Result<()>

Emit function locals

the function parameters.

function body locals.

NOTE: we don’t care about the origin offset of the locals. bcz we will serialize the locals to an index map anyway.

pub fn emit_operators( &mut self, ops: &mut OperatorsReader<'_>, validator: &mut FuncValidator<ValidatorResources> ) -> Result<()>

Emit function operators

pub fn finish(self, jump_table: &mut JumpTable, pc: u16) -> Result<Buffer>

Finish code generation.

impl CodeGen

pub fn _call_indirect( &mut self, _type_index: u32, _table_index: u32, _table_byte: u8 ) -> Result<()>

The call indirect instruction calls a function indirectly through an operand indexing into a table.

pub fn _call(&mut self, index: u32) -> Result<()>

The call instruction calls a function specified by its index.

fn call_internal(&mut self, index: u32) -> Result<()>

Call internal functions

fn call_imported(&mut self, index: u32) -> Result<()>

Call imported functions

impl CodeGen

pub fn _if(&mut self, blockty: BlockType) -> Result<()>

The beginning of an if construct with an implicit block.

pub fn _block(&mut self, blockty: BlockType) -> Result<()>

The begeinning of a block construct. A sequence of instructions with a label at the end.

pub fn _loop(&mut self, blockty: BlockType) -> Result<()>

A block with a label which may be used to form loops.

pub fn _else(&mut self) -> Result<()>

Marks an else block of an if.

pub fn _select(&mut self) -> Result<()>

The select instruction selects one of its first two operands based on whether its third oprand is zero or not.

STACK: [cond, val2, val1] -> [val1] if cond is non-zero, [val2] otherwise.

pub fn _br(&mut self, _depth: u32) -> Result<()>

Branch to a given label in an enclosing construct.

Performs an unconditional branch.

pub fn _br_if(&mut self, depth: u32) -> Result<()>

Performs a conditional branch if i32 is non-zero.

Conditional branch to a given label in an enclosing construct.

pub fn _br_table(&mut self, _table: BrTable<'_>) -> Result<()>

A jump table which jumps to a label in an enclosing construct.

Performs an indirect branch through an operand indexing into the label vector that is an immediate to the instruction, or to the default target if the operand is out of bounds.

pub fn _end(&mut self) -> Result<()>

Handle the end of instructions for different situations.

TODO: (#28)

End of control flow operators.

End of function.

End of program.

pub fn _unreachable(&mut self) -> Result<()>

Mark as invalid for now.

TODO: recheck this implementation, if it is okay, provide more docs.

pub fn _nop(&mut self) -> Result<()>

Perform nothing in EVM bytecode.

impl CodeGen

pub(crate) fn handle_return(&mut self) -> Result<()>

Handle the end of the function.

pub(crate) fn handle_call_return(&mut self) -> Result<()>

Handle the return of a call.

pub(crate) fn handle_frame_popping( &mut self, frame: ControlStackFrame ) -> Result<()>

Handle the popping of a frame.

TODO: validate stack IO for all frames (#59)

pub(crate) fn handle_jumpdest(&mut self, original_pc: u16) -> Result<()>

Handle jumpdest.

impl CodeGen

pub fn _local_get(&mut self, local_index: u32) -> Result<()>

This instruction gets the value of a variable.

pub fn _local_set(&mut self, local_index: u32) -> Result<()>

This instruction sets the value of a variable.

pub fn _local_tee(&mut self, index: u32) -> Result<()>

This _local_tee is like _local_set, but it also returns the value on the stack.

pub fn _global_get(&mut self, _: u32) -> Result<()>

This instruction gets the value of a variable.

pub fn _global_set(&mut self, _: u32) -> Result<()>

This instruction sets the value of a variable.

fn _local_get_calldata(&mut self, local_index: usize) -> Result<()>

Local get from calldata.

fn _local_get_var(&mut self, local_index: usize) -> Result<()>

Local get for variables.

impl CodeGen

fn log_data(&mut self) -> Result<(i32, i32)>

Parse log data from the bytecode.

pub fn log(&mut self, count: usize) -> Result<()>

Log a message with topics.