Struct zingen::MacroAssembler

pub struct MacroAssembler {
    pub(crate) asm: Assembler,
}

EVM MacroAssembler.

Fields
asm: Assembler

Low level assembler.

Implementations

impl MacroAssembler

pub fn _ge(&mut self) -> Result<()>

Greater than or equal comparison.

a b ge -> a b-1 gt(lt)

Using lt due to order of stack.

pub fn _sge(&mut self) -> Result<()>

Greater than or equal comparison.

a b sge -> a b-1 sgt(slt)

Using lt due to order of stack.

pub fn _sle(&mut self) -> Result<()>

Greater than or equal comparison.

a b sge -> a b-1 sgt(slt)

Using lt due to order of stack.

pub fn _le(&mut self) -> Result<()>

Greater than or equal comparison.

a b le -> a b-1 lt(gt)

Using gt due to order of stack.

pub fn _sgt(&mut self) -> Result<()>

Greater than and equal comparison.

Using slt due to order of stack.

pub fn _gt(&mut self) -> Result<()>

Greater than comparison.

Using lt due to order of stack.

pub fn _lt(&mut self) -> Result<()>

less than comparison.

Using gt due to order of stack.

pub fn _slt(&mut self) -> Result<()>

less than or equal comparison.

Using gt due to order of stack.

pub fn _ne(&mut self) -> Result<()>

Sign-agnostic compare unequal.

pub fn _eqz(&mut self) -> Result<()>

Simple not operator

impl MacroAssembler

pub fn _max(&mut self) -> Result<()>

Maximum of two values

pub fn _min(&mut self) -> Result<()>

Minium of two values

pub fn _ceil(&mut self) -> Result<()>

Ceiling operator

pub fn _floor(&mut self) -> Result<()>

Floor operator

pub fn _nearest(&mut self) -> Result<()>

Round to nearest integer, ties to even.

pub fn _sqrt(&mut self) -> Result<()>

Square root

pub fn _abs(&mut self) -> Result<()>

Absolute value

pub fn _neg(&mut self) -> Result<()>

Negation

pub fn _copysign(&mut self) -> Result<()>

If z1 and z2 have the same sign, return z1, otherwise return z1 with negated sign.

pub fn _convert_i32_s(&mut self) -> Result<()>

Convert a signed 32-bit integer to a (32-bit/64-bit) float

pub fn _convert_i32_u(&mut self) -> Result<()>

Convert an unsigned 32-bit integer to a (32-bit/64-bit) float

pub fn _convert_i64_s(&mut self) -> Result<()>

Convert a signed 32-bit integer to a (32-bit/64-bit) float

pub fn _convert_i64_u(&mut self) -> Result<()>

Convert a unsigned 32-bit integer to a (32-bit/64-bit) float

pub fn _trunc(&mut self) -> Result<()>

Round to nearest integer towards zero

impl MacroAssembler

pub fn _sub(&mut self) -> Result<()>

Sub two numbers.

pub fn _i32_const(&mut self, value: i32) -> Result<()>

Push a 32-bit integer value on the stack.

pub fn _i64_const(&mut self, _value: i64) -> Result<()>

Push a 64-bit integer value on the stack.

pub fn _f32_const(&mut self, _value: Ieee32) -> Result<()>

Push a 32-bit float value on the stack.

pub fn _f64_const(&mut self, _value: Ieee64) -> Result<()>

Push a 64-bit float value on the stack.

pub fn _i32_wrap_i64(&mut self) -> Result<()>

wrap a 64-bit integer to a 32-bit integer.

pub fn _i64_extend_i32_s(&mut self) -> Result<()>

Extend a signed 32-bit integer to a 64-bit integer.

pub fn _i64_extend_i32_u(&mut self) -> Result<()>

Extend an unsigned 32-bit integer to a 64-bit integer.

pub fn _f32_demote_f64(&mut self) -> Result<()>

Truncate a 64-bit float to a signed 32-bit integer.

pub fn _f64_promote_f32(&mut self) -> Result<()>

Truncate a 64-bit float to an unsigned 32-bit integer.

pub fn _i32_reinterpret_f32(&mut self) -> Result<()>

Convert a signed 32-bit integer to a 32-bit float.

pub fn _i64_reinterpret_f64(&mut self) -> Result<()>

Convert a signed 64-bit integer to a 64-bit float.

pub fn _f32_reinterpret_i32(&mut self) -> Result<()>

Convert a 32-bit float to a signed 32-bit integer.

pub fn _f64_reinterpret_i64(&mut self) -> Result<()>

Convert a 64-bit float to a signed 64-bit integer.

pub fn _rotl(&mut self) -> Result<()>

sign-agnostic rotate left

Return the result of rotating i1 left by k bits.

pub fn _rotr(&mut self) -> Result<()>

sign-agnostic rotate right

Return the result of rotating i1 right by k bits.

pub fn _clz(&mut self) -> Result<()>

sign-agnostic count leading zero bits

Return the number of leading zero bits in i, all zero bits are considered leading if the value is zero.

pub fn _ctz(&mut self) -> Result<()>

sign-agnostic count leading zero bits

Return the number of leading zero bits in i, all zero bits are considered trailing if the value is zero.

pub fn _popcnt(&mut self) -> Result<()>

sign-agnostic count number of one bits

Return the count of no zero bits in i.

pub fn _trunc_f32(&mut self) -> Result<()>

Truncate a 32-bit float to an integer

pub fn _trunc_f64(&mut self) -> Result<()>

Truncate a 64-bit float to an integer

impl MacroAssembler

pub(crate) fn _load(&mut self) -> Result<()>

Load n bytes to extend self as another number type.

Just for adapting the WASM instructions, this method makes no sense for EVM since all of the numbers as U256.

pub(crate) fn _load8(&mut self) -> Result<()>

Load 1 byte to extend self as another number type.

Just for adapting the WASM instructions, this method makes no sense for EVM since all of the numbers as U256.

pub(crate) fn _load16(&mut self) -> Result<()>

Load 2 bytes to extend self as another number type.

Just for adapting the WASM instructions, this method makes no sense for EVM since all of the numbers as U256.

pub(crate) fn _load32(&mut self) -> Result<()>

Load 4 bytes to extend self as another number type.

Just for adapting the WASM instructions, this method makes no sense for EVM since all of the numbers as U256.

pub fn _store(&mut self) -> Result<()>

Store n bytes in memory.

pub fn _store8(&mut self) -> Result<()>

Wrap self to i8 and store 1 byte

pub fn _store16(&mut self) -> Result<()>

Wrap self to i16 and store 2 bytes

pub fn _store32(&mut self) -> Result<()>

Wrap self to i32 and store 4 bytes

pub fn _memory_size(&mut self, _: u32, _: u8) -> Result<()>

The memory size instruction returns the current size of memory.

pub fn _memory_grow(&mut self, _: u32, _: u8) -> Result<()>

The memory grow instruction grows memory by a given delta and returns the previous size, or -1 if enough memory cannot be allocated.

impl MacroAssembler

pub fn _drop(&mut self) -> Result<()>

The drop instruction simply throw away a single operand.

impl MacroAssembler

pub fn memory_write(&mut self, ty: impl Type) -> Result<usize>

Store data in memory with at current memory byte pointer.

pub fn memory_write_at(&mut self, offset: &[u8]) -> Result<()>

Store data in memory at offset.

Returns the size in the lowest significant bytes.

pub fn pc_offset(&self) -> u16

Get the current program counter offset.

pub fn push(&mut self, bytes: &[u8]) -> Result<()>

Place n bytes on stack.

pub fn mp_offset<F>(&self, f: F) -> Result<SmallVec<[u8; 8]>>where F: Fn(usize) -> Result<usize>,

Get byte offset of the memory pointer.

pub fn sp(&self) -> u8

Get the stack pointer.

pub fn swap(&mut self, index: u8) -> Result<()>

Swap memory by target index.

pub fn dup(&mut self, index: u8) -> Result<()>

Duplicate stack item by target index.

pub fn shift_pc(&mut self, count: u8, from_top: bool) -> Result<()>

Shift the program counter to the bottom or the top of the parameters. This is used by the callee function for jumping back to the caller function.

pub fn _return(&mut self) -> Result<()>

Return zero or more values from the function.

The return instruction is a shortcut for an unconditional branch to the outermost block, which implicitly is the body of the current function.

NOTE: This return could be different from the return in the EVM.