zingen
Module visitor

This module is the central place for machine code emission.

It defines an implementation of wasmparser’s Visitor trait for Function; which defines a visitor per op-code, which validates and dispatches to the corresponding machine code emitter.

Modules

call 🔒

call instructions

control 🔒

Control flow visitors

local 🔒

Local instructions

log 🔒

System instructions

Macros

impl_visit_operator 🔒

A macro to define unsupported WebAssembly operators.

map_wasm_operators 🔒

Implement arithmetic operators for types.