Crate zingen

Code generation library for zink.

Re-exports

pub use crate::abi::ToLSBytes;

pub use crate::abi::Type;

Modules

abi

WASM ABI

asm 🔒

Low level assembler implementation for EVM.

backtrace 🔒

Backtrace support for the code generation.

code 🔒

Table for the code section.

codegen 🔒

Code generation implementation.

constructor 🔒

Contract constructor.

control 🔒

Data structures for control flow emission.

data 🔒

Dataset in code generation

dispatcher 🔒

Code generator for EVM dispatcher.

export 🔒

Exports in code generation.

func 🔒

Function handler

import 🔒

Built-in functions for EVM

jump 🔒

Jump table implementation.

local 🔒

WASM local slot.

masm 🔒

MacroAssembler used by the code generation.

result 🔒

Codegen results

validator 🔒

Pre-visitor for parsing WASM.

visitor 🔒

This module is the central place for machine code emission.

Structs

Assembler

Low level assembler implementation for EVM.

Code

Code section for EVM.

CodeGen

The code generation abstraction.

Constructor

Contract constructor.

ControlStack

The control stack.

ControlStackFrame

Holds the necessary metadata to support the smission of control flow instructions.

DataSet

Data section conversion

Dispatcher

Code generator for EVM dispatcher.

Exports

WASM exports

Function

Function with validator.

Functions

Functions with indexes.

Imports

Imported functions

JumpTable

Jump table implementation.

LocalSlot

A local slot.

Locals

Solidity’s implementation uses 16 slots for locals. ref: https://docs.soliditylang.org/en/v0.8.20/internals/optimizer.html#stackcompressor

MacroAssembler

EVM MacroAssembler.

Enums

ControlStackFrameType

The type of the control stack frame.

Error

Codegen error

Func

EVM built-in function.

Constants

BUFFER_LIMIT

Maximum size of a evm bytecode in bytes.

Type Aliases

Buffer

Code generation buffer.

Result

Codegen result