zingen
Module wasm

WASM related primitives.

Modules

abi 🔒

WASM ABI

data 🔒

Dataset in code generation

func 🔒

Function handler

host 🔒

Host functions

Macros

impl_deref 🔒

Structs

Data

Data section conversion

Env

A struct that holds the environment wasm module.

Exports

WASM export section

Function

Function with validator.

Functions

WASM Functions by indexes.

Imports

WASM import section

Enums

HostFunc

EVM built-in function.

Traits

ToLSBytes

Get the offset of this type in the lowest significant bytes.

Type

WASM type size for the stack representation of EVM.